requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
